requirements.txt
cloud-volume
python-dotenv
slackclient
click
pika
google-cloud-storage